Rename swv.js to validate.js
And move validation functions to it.

#749
